+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ### Depracated
 
+## v0.0.2 - 2022/4/2
+
+### New Features
+
+* Switched to TypeScript
+* Ethereum address validation added to check
+
 ## v0.0.1 - 2022/4/2
 
 * Initial Alpha release